+++ src/main/java/edu/temple/cis/paystation/RateStrategyOmega.java
@@ -2,7 +2,6 @@ package edu.temple.cis.paystation;
 
 import java.time.DayOfWeek;
 import java.time.LocalDate;
-import java.util.Random;
 
 /**
  * Implementation of RateStrategy for Omega town.
@@ -19,7 +18,7 @@ public class RateStrategyOmega implements RateStrategy {
     private final RateStrategy rsA = new RateStrategyAlpha();
 
     @Override
-    public double calculateTime(int amount) {
+    public int calculateTime(int amount) {
         //date = the systems current date, then dayOfTheWeek number is the systems current date turned into a number representing day of the week
         LocalDate date = LocalDate.now();
         int dayOfTheWeekNumber = getDayNumber(date);
